Fix change, delete and the summary page behaviour.
Add the `params` to make change and delete work.
Amend the loop.js behaviour to look for `loopSections` rather than sections
so that adding loop fields to the summary doesn't require declaring all
sections in the corresponding *-summary.js file.
Finally, move the SetConfirmStep behaviour to the loop step - introduces a
slight inefficiency but it means that it is always set, even if the app is
redeployed when someone is part way through a flow.