Amend the back link behaviour on loop pages.
New behaviour is:
If first entry page of the first item then allow superclass to handle.
If on the summary screen then allow superclass to handle.
If in a data entry flow return to previous data entry screen.
If on the first page of a data entry flow then return to summary screen.