Continue on edit
Due to the nature of the form flow (branches based on previous
answers, some separated by several pages), and the lack of support
for continueOnEdit as a function in HOF at the present time we have
elected to simply repeat the remainder of the form whenever a change
happens for the time being.
For clarity the root of the issue is that when editing e.g. brp details
it's not currently possible to determine whether we now need to show
the remainder of the single / joint journey or not